Add files via upload
This code will play four different Christmas songs with a passive buzzer. Eight LEDs will blink with the beat of the music and an extra LED blinks every time the buzzer makes a sound. Can switch between tracks if using a button or switch connected to analog pin A0 on the Arduino.